Changed Ucm to be based upon Odict and changed return of 'off'
The old Ucm class was inherited from a list to preserve the header item order
but this made header access inconvenient. I have now switched to my Odict
'ordered dictionary' class. I also changed the type of the pixel offset
attribute so that a specific window of a specific CCD will give a 2-element
tuple with (llx,lly). Finally I spotted a bug in the write routine which I
have fixed.